+++ src/api/v2/shelter-admin/index.ts
@@ -2,7 +2,7 @@ import { POST } from "@/api/axios";
 import * as ROUTES from "./Routes.const";
 import { AdminDataType } from "@/components/Layout/CreateAdminForm";
 import { generateSplitUrl } from "@/api/utils.const";
-import { CreateAdminResponseType } from "@/api/v1/shelter-admin/type";
+import { CreateAdminResponseType } from "@/api/v2/shelter-admin/type";
 
 export function createAdminAccount(adminData: AdminDataType): Promise<CreateAdminResponseType> {
     return POST({ url: ROUTES.CREATE_ADMIN_ACCOUNT, data: adminData })
