+++ src/api/v2/shelter-admin/type.d.ts
@@ -1,0 +1,8 @@
+export type AdminId = number
+
+export interface CreateAdminErrorType {
+  errorCode: string
+  description: string
+}
+
+export type CreateAdminResponseType = AdminId | CreateAdminErrorType
